Extracted common interface between V1 and V2
