+++ polaris/experimental/_dataset_v2.py
@@ -1,0 +1,5 @@
+from polaris.dataset._base import BaseDataset
+
+
+class DatasetV2(BaseDataset):
+    """Dataset subclass for Polaris"""
